Implement a mechanism to convert from local to global coordinates.
This is needed to convert the local coordinates retrieved by Geant4
during the overlap check activity, into global coordinates
that can be then written out to the json file that will contain the
clashes report. These global coordinates can then be read in in gmex
and visualized, to help debugging them.