Fix to the local to global coordinates transformation for clashing points
This fixes in particular the conversion from local to global coordinates
of clashing points with the mother volume. The reported local clashing point
is local to the mother and not to the volume. With this fix one step in the conversion
chain is skipped, if the clash is with the mother volume, since the point is already
expressed in the mother volume coordinate system.

Also this should fix the conversion from local to global coordinates of clashes with a sister volume.
In this case the point of clashing is in the coordinate system of the sister. With the transformation
added we convert this point into the coordinate system of the analized volume.