Fix build when nlohmann_json is not CMake-installed
When nlohmann_json is installed in other ways rather
than by using the CMake setup included within the
package, no CMake config file is installed for
nlohmann_json; and so, it is not possible to use
find_package( nlohmann_json ) in the CMake setup of
our package. This commit fixes the build for the case,
for example, when nohmann_json is installed
by simply copying the single-header file into a system
include folder as '/usr/local/include'.