Ass a new case to the GXClashPointSystem, related to invalid solids.
A solid is considered not correctly defined, so invalid, when it's not possible
to generate points on its surface. In this case gmclash writes an entry in the
json report, writing as coordinates the global coordinates of the center
of the invalid volume. The field distance is filled with a generic non-valid value: -999
These points are visualized in gmex in color folly, under the checkbox InvalidSolid