Add mutex to protect access to PhysVol 'daughters' vector.
The GeoPhysVol::add() method adds a child to a GeoPhysVol.
The underlaying method which increases the number of children (RCBase::Ref()) is thread-safe
because the counter is implemented with a std::atomic. But the vector storing the children
of a GeoPhysVol is not thread-safe because the access to its members is not protected in any way.
I added a mutex and locks to all methods writing to or reading from the vector.