Bug-fix GDMLtoGM, that caused segfault with 'auxiliary'
Auxiliary is not implemented in GeoModel, so there is no handler for it. GetHandler() method returns a null pointer
that it is used later in the code, causing the segfault. This commit protects the code using this pointer with a check
on the pointer itself. This commit should solve all the cases of not defined handler for volumes.