Introduced a local copy of LArCustomShape to be used internally by LArPlugin
The current implementation is for demonstration purposes only!

The main goal of it to demonstrate that we can build LAr GeoModel in standalone
mode, i.e. with no dependency on Athena, and in particular with no access to
the Geometry DB via RDBAccessSvc.

For latter we implemented a new structure LArWheelCalculatorParameters which holds
all parameters which can be filled from either XML input, or Geometry DB, or any other
source. The caller code (either LAr Plugin, or LAr Athena Factory) is supposed to create
and fill this structure, and the pass it to the constructor of either LArCustomShape or
LArWheelCalculator.

The current implementation is broken is several places:

 * Instead of doing a proper migration from CLHEP to Eigen, the real code was replaced
by dummies
 * The dependency on ROOT was just dropped using brute force approach

Both of the above problems need proper fixing!