Rationalised the code in SetupJSON.cmake a bit.
The module will now create an nlohmann_json::nlohmann_json target
no matter what. Even if the external is not found, it will still
set up a dummy target. This way the rest of the GeoModel code does
not need to check constantly whether an nlohmann_json target is
available or not.