GeoModelKernel: Compile GeoXF.cxx with optimization, even in debug builds.
GeoXF.cxx makes heavy use of Eigen.  In debug builds it ends up being
much, much slower than in optimized builds, to the point where where it
makes it difficult to run things in dbg.  Tweak the build so that we compile
GeoXF.cxx with optimization, even in dbg builds.  We also need to use the
flatten attribute on the functions that use Eigen, so ensure that the Eigen
operations are completely inlined.

This was motivated by the fact that numerous tests are timing out in the dbg
build.  This change alone reduced the time required by one test from ~600s
to ~400s.