Update the PYTHIA based primary generator with proper MT seeding.
Originally, a fixed (1234) seeding was used for seeding the RNG
of Pythia. In case of using Geant4 MT with multiple worker threads,
this results in generating exactly the same set of events for each
of the threads. This behaviour is kept now only for performance
mode runs (-p) while the worker thread generators are seeded now
using the event ID ortherwise. This results in a reproducible
set of events for the complete run while the individual worker thraeds
simulate distinct subsets of this.