Cleanup of the fullSimLight application
A significant clean up has been done on the application, especially
regarding the user actions. These were still identical to that I
developed for the GeantV project in the fullCMS Geant4 application
(that fullSimLight is based on). While the goal of that application
was to demonstrate even some more advanced user action features,
these are completely unnecessary for fullSimLight and causing only
limitations. So the user actions has been reduced, made simpler and
cleaner. Only a minimum set of infomation is collected now (only if
executed in non-performance mode) such as the mean energy deposit,
mean number of charged/neutral steps, mean number of secondary electron,
gamma and positron per event.

An additional option, for collecting the above infomation only
restricted to a given detector region, will be added later when
activating detector regions will already be implemented.

Note, that the same user actions can now be used both with the
particle gun and the Pythia primary event generator options.