WoodcockProcess, for Woodcock tracking of gamma photons in a given detector region, has been implemented.
The WoodcockProcess has been implemented for utilising Woodcock tracking of gamma
photons in a given detector region. The Woodcock process is a specialised version
of the Gamma-general process: Woodcock tracking is applied within the given
detector region (specified by the user at the construction of the WoodcockProcess)
while the WoodcockProcess falls into the Gamma-general one otherwise. It is also
the case if no detector region with the specified name can be found in the detector
or the kinetic energy of the gamma drops below a certain energy. This energy limit
can be set at construction time (0 by default).

Since the WoodcockProcess would require a very recent version of the base
G4GammaGeneralProcess (the one that will be available in Geant4-11.0), a local
version of that process (a very similar to that in G4.11.0) is also be given in
the form of the GammaGeneralProcess. This is used as a base for the current
WoodcockProcess as long as a more recent version of Geant4 will be utilised by
the project.