Adding the local extra physics constructor for gamma-nuclear.
Gamma particles has extra, i.e. non-EM processes (such as gamma nuclear) that
are usually added in the G4EmExtraPhysics constructor. This is done properly even
in the case of using Gamma-general process which means that these extra processes
are added to the G4GammaGeneralProcess instead of the gamma particle whenever
Gamma-general process is used. Since in our case we do not use the G4GammaGeneralProcess
as base class for the WoodcockProcess but the GammaGeneralProcess, we also need
need a local version of this G4EmExtraPhysics that will add the extra (gamma-
nuclear) process to the GammaGeneralProcess. This local EmExtraPhysics constructor
will be suitable for Geant4.10.6.p03.