Change behavior of I/O helper functions, add a new 'getReader'helper.
The 'loadDB' helper now returns a nullptr if the path of the .db file is
wrong or not readable. It is up to the client to check teh pointer and
act accordingly.

The new 'getReaderDB' helper returns a GeoModelIO::ReadGeoModel object,
and exits if the path of the input .db file is wrong or not readable. I
assume lower-level needs from clients that want to access the reader to
get infdo from the database, so I think an exit function is OK here, if
the .db file is not found.