Add constness to methods loading .db and returning 'world'.
Sarka Todorova correctly noted that the public loaders should return
a const 'world' volume, so a const pointer to the GeoVPhysVol.

I added constness to both the GeoModelRead::buildGeoModel() and to the
new loadDB() helper methods.

I then remove the constness in the visualization code, internally,
otherwise it's more complicated to manipulate the geo tree.