Modification of the way in which Transfunctions are persistified. Step 1: Genfunctions
Before:  Floating point data which is part of the description of
the stored functions is written into the string, the same string
used to define the function.

After:  Floating point data is stored in an auxilliary array that
must be used together with the string when the function is restored.
The array is implemented as an std::deque. It functions like a
FIFO.

Step 2: Transfunctions

(Not yet available at this stage)