Modification of the way in which Transfunctins are persistified (conclusion)
Step 2:

Before:  Floating point data which is part of the description of
the stored TransFunctions is written into the string, the same string
used to define the function.

After:  Floating point data is stored in an auxilliary array that
must be used together with the string when the Transfunction is restored.
The array is implemented as an std::deque. It functions like a
FIFO. This array also includes the floating point data from any
Genfunctions associated with the Transfunctions; that data was
already available after step 1 was committed.