Fix safety tests
Until you build the GeoModel tree, the caches are not created yet; however, the low-level methods can access the DB directly. So, the existance of a table should be verified at the DB level, not at cache level; otherwise the test fails when calling methods before building the GeoModel tree.